17.04.2023 add calendar booking

+++ src/main/java/com/rzaglada1/booking/controllers/HouseController.java
@@ -3,8 +3,10 @@ package com.rzaglada1.booking.controllers;
 
 import com.rzaglada1.booking.models.Address;
 import com.rzaglada1.booking.models.House;
+import com.rzaglada1.booking.models.OrderHistory;
 import com.rzaglada1.booking.models.enams.Role;
 import com.rzaglada1.booking.services.HouseService;
+import com.rzaglada1.booking.services.OrderHistoryService;
 import com.rzaglada1.booking.services.UserService;
 import com.rzaglada1.booking.services.WishService;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +19,7 @@ import java.io.IOException;
 import java.security.Principal;
 
 
+
 @Controller
 @RequestMapping(value = "/houses")
 @RequiredArgsConstructor
@@ -24,6 +27,7 @@ public class HouseController {
     private final HouseService houseService;
     private final UserService userService;
     private final WishService wishService;
+    private final OrderHistoryService orderHistoryService;
 
 
     // request form all  house
@@ -31,7 +35,9 @@ public class HouseController {
     public String houseAll(Model model, Principal principal) {
         if (principal != null && userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)) {
             model.addAttribute("houses", houseService.getAll());
-        } else {
+            model.addAttribute("admin", "admin");
+            model.addAttribute("user", userService.getUserByPrincipal(principal));
+            } else {
             model.addAttribute("houses", houseService.getHouseByUser(userService.getUserByPrincipal(principal)));
         }
 
@@ -40,9 +46,13 @@ public class HouseController {
 
     // request form new  house
     @GetMapping("/new")
-    public String houseNew(Model model) {
+    public String houseNew(Model model, Principal principal) {
         model.addAttribute("house", null);
         model.addAttribute("address", new Address());
+        model.addAttribute("user", userService.getUserByPrincipal(principal));
+        if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)) {
+            model.addAttribute("admin", "admin");
+        }
 
         return "house/house_form";
     }
@@ -108,21 +118,30 @@ public class HouseController {
     @GetMapping("/{id}/edit")
     public String houseEdit(@PathVariable Long id, Model model, Principal principal) {
 
-        System.out.println("11111 " + houseService.getById(id).get().getPrice());
         if (principal != null && houseService.getById(id).isPresent()) {
             if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)
                     || houseService.getHouseById(id).orElseThrow().getUser().equals(userService.getUserByPrincipal(principal))
             ) {
                 model.addAttribute("house", houseService.getById(id).get());
                 model.addAttribute("address", houseService.getById(id).get().getAddress());
+                model.addAttribute("user", userService.getUserByPrincipal(principal));
+            }
+            if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)) {
+                model.addAttribute("admin", "admin");
             }
         }
         return "house/house_form";
     }
 
     @GetMapping("/{id}/delete")
-    public String houseDelete(@PathVariable("id") Long id) {
-        houseService.deleteById(id);
+    public String houseDelete(@PathVariable("id") Long id, Principal principal) {
+        if (principal != null && houseService.getById(id).isPresent()) {
+            if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)
+                    || houseService.getHouseById(id).orElseThrow().getUser().equals(userService.getUserByPrincipal(principal))
+            ) {
+                houseService.deleteById(id);
+            }
+        }
 
         return "redirect:/houses";
     }
@@ -130,25 +149,67 @@ public class HouseController {
     @GetMapping("/{houseId}/detail")
     public String houseDetail(@PathVariable("houseId") Long houseId, Model model, Principal principal) {
         if (houseService.getById(houseId).isPresent()) {
-            model.addAttribute("house", houseService.getById(houseId).get());
-            model.addAttribute("feedbacks", houseService.getById(houseId).get().getFeedbackList());
-            model.addAttribute("user", houseService.getById(houseId).get().getUser());
-            model.addAttribute("user", userService.getUserByPrincipal(principal));
+            House house = houseService.getById(houseId).get();
+            if (!orderHistoryService.findOrdersByHouseForFree(house).isEmpty()) {
+                model.addAttribute("orders", orderHistoryService.findOrdersByHouseForFree(house));
+            }
+            model.addAttribute("house", house);
+            model.addAttribute("feedbacks", house.getFeedbackList());
+
             if (principal != null) {
+                if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)) {
+                    model.addAttribute("admin", "admin");
+                }
+                model.addAttribute("user", userService.getUserByPrincipal(principal));
                 model.addAttribute("isWishList", wishService.existsByHouseIdAndUser(houseId, principal));
             }
+
+
 
         }
         return "house/house_detail";
     }
 
-    @PostMapping("/{houseId}/booking")
-    public String hoseBooking (@PathVariable("houseId") long houseId, Principal principal) {
+    @PostMapping("/{houseId}/prebooking")
+    public String hosePreBooking(
+            @PathVariable("houseId") long houseId,
+            OrderHistory orderHistory,
+            Model model,
+            Principal principal) {
 
-        return " ";
+        if (houseService.isDateFree(orderHistory, houseId)) {
+
+            House house = houseService.getHouseById(houseId).get();
+            orderHistory.setHouse(house);
+            model.addAttribute("house", house);
+            model.addAttribute("order", orderHistory);
+
+            return "/house/house_booking";
+        } else {
+            model.addAttribute("message", "На вказаний період вже заброньовано. Спробуйте інші дати");
+            return "/message";
+        }
     }
 
+    @PostMapping("/{houseId}/booking")
+    public String hoseBooking(
+            @PathVariable("houseId") long houseId,
+            OrderHistory orderHistory,
+            Model model,
+            Principal principal) {
 
+        if (houseService.getHouseById(houseId).isPresent() && principal != null) {
+            House house = houseService.getHouseById(houseId).get();
+            orderHistory.setHouse(house);
+            orderHistoryService.saveToBase(orderHistory, houseId, principal);
+            model.addAttribute("message", "Заброньовано");
+        } else {
+            model.addAttribute("message", "Щось пішло не так. Спробуйте знову.");
+        }
+
+        return "/message";
+
+    }
 
 
 }
