+++ src/main/java/com/rzaglada1/booking/controllers/WishController.java
@@ -1,6 +1,8 @@
 package com.rzaglada1.booking.controllers;
 
 import com.rzaglada1.booking.models.Wish;
+import com.rzaglada1.booking.models.enams.Role;
+import com.rzaglada1.booking.services.UserService;
 import com.rzaglada1.booking.services.WishService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -14,25 +16,39 @@ import java.security.Principal;
 @RequiredArgsConstructor
 public class WishController {
     private final WishService wishService;
+    private final UserService userService;
 
     @GetMapping()
     public String wishes (Model model, Principal principal) {
-        model.addAttribute("wishes", wishService.getWishByUser(principal));
+        if (userService.getUserByPrincipal(principal).getEmail() != null) {
+            model.addAttribute("wishes", wishService.getWishByUser(principal));
+            model.addAttribute("user", userService.getUserByPrincipal(principal));
+            if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)) {
+                model.addAttribute("admin", "admin");
+            }
+        }
+
 
         return "wishes/wishes_list";
     }
     @GetMapping("/{houseId}/delete")
     public String wishes (Model model, @PathVariable long houseId, Principal principal) {
-        wishService.deleteFromBase(houseId,principal);
-        model.addAttribute("wishes", wishService.getWishByUser(principal));
-
+        if (userService.getUserByPrincipal(principal) != null) {
+            wishService.deleteFromBase(houseId,principal);
+            model.addAttribute("wishes", wishService.getWishByUser(principal));
+            model.addAttribute("user", userService.getUserByPrincipal(principal));
+            if (userService.getUserByPrincipal(principal).getRoles().contains(Role.ROLE_ADMIN)) {
+                model.addAttribute("admin", "admin");
+            }
+        }
         return "wishes/wishes_list";
     }
 
     @PostMapping("/{idHouse}/new")
     public String houseToWish(Wish wish, @PathVariable long idHouse, Principal principal) {
-        wishService.saveToBase(wish, idHouse, principal);
-
+        if (userService.getUserByPrincipal(principal) != null) {
+            wishService.saveToBase(wish, idHouse, principal);
+        }
        // return "redirect:/";
         return "redirect:/houses/" + idHouse + "/detail";
     }
